requirements.txt
pytorch
numpy
matplotlib
datasets
transformers
tqdm
pytest
pytest-cov
